Cleaned up code for server integration; Added user creation

+++ VeloToApp/CoreData/ActionCardsCoreDataWrapper.swift
@@ -34,7 +34,7 @@ struct ActionCardsCoreDataWrapper: CoreDataWrapper {
         return CoreDataHelper.retrieveAll(name: entityName)
     }
     
-    static func retrieveAllForAthleteID(id athlete: Int32) -> [ActionCard]? {
+    static func retrieveAllForAthleteID(id athlete: Int64) -> [ActionCard]? {
         return CoreDataHelper.retrieveAll(name: entityName, with: "athlete.id == %@", arguments: [athlete])
     }
